+++ lnmp-docker.sh
@@ -4,10 +4,23 @@ ENV=$1
 ARCH=`uname -m`
 OS=`uname -s`
 
+# 获取正确版本号
+
+. env/lnmp.env
+
+if [ OS = "Darwin" ];then
+  # 将以什么开头的行替换为新内容
+  sed -i '' "s/^KHS1994_LNMP_DOCKER_VERSION.*/KHS1994_LNMP_DOCKER_VERSION=${KHS1994_LNMP_DOCKER_VERSION}/g" .env
+else
+  sed -i "s/^KHS1994_LNMP_DOCKER_VERSION.*/KHS1994_LNMP_DOCKER_VERSION=${KHS1994_LNMP_DOCKER_VERSION}/g" .env
+fi
+
+
 # 不支持信息
 
 NOTSUPPORT(){
-  echo -e "\033[32mINFO\033[0m `uname -s` ${ARCH} 暂不支持\n"
+  echo -e "\033[32mINFO\033[0m  Not Support `uname -s` ${ARCH}\n"
+  exit 1
 }
 
 # 创建日志文件
@@ -129,19 +142,6 @@ function env_status(){
 fi
 }
 
-# 初始化 Gogs
-
-function gogs_init(){
-  # .env.example to .env
-  if [ -f app.prod.ini ];then
-    echo -e "\033[32mINFO\033[0m  app.prod.ini 文件已存在\n"
-  else
-    echo -e "\033[31mINFO\033[0m  app.prod.ini 文件不存在\n"
-    cp app.ini app.prod.ini
-    # exit 1
-fi
-}
-
 # 初始化
 
 function init {
@@ -229,6 +229,7 @@ main() {
   echo -e "\n\033[32mINFO\033[0m  ARCH is ${OS} ${ARCH}\n"
   env_status
   . .env
+  . env/.env
   case $1 in
 
   init )
@@ -287,18 +288,6 @@ main() {
          -f docker-compose.yml \
          -f docker-compose.prod.yml \
          config
-    ;;
-
-  ci )
-    cd ci
-    if [ -f ".env" -a -f "update.js" ];then
-      docker-compose up -d
-    else
-      echo -e "\033[31mINFO\033[0m  .env .update.js update.sh 文件不存在 \n"
-      cp update.example.sh update.sh \
-         && cp update.example.js update.js \
-         && cp .env.example .env
-    fi
     ;;
 
   development-build )
@@ -435,11 +424,6 @@ services:
     docker-compose down --remove-orphans
     ;;
 
-  gogs-init )
-    cd config/gogs
-    gogs_init
-    ;;
-
   * )
   echo  -e "
 Docker-LNMP CLI ${KHS1994_LNMP_DOCKER_VERSION}
@@ -449,12 +433,10 @@ Usage: ./docker-lnmp COMMAND
 Commands:
   backup               Backup MySQL databases
   cleanup              Cleanup log files
-  ci                   Use CI/CD in Production
   composer             Use PHP Package Management composer
   development          Use LNMP in Development(Support x86_64 arm32v7 arm64v8)
   development-config   Validate and view the Development(with build images) Compose file
   development-build    Use LNMP in Development With Build images(Support x86_64)
-  gogs-init            Initialize Gogs
   help                 Display this help message
   laravel              Create a new Laravel application
   laravel-artisan      Use Laravel CLI artisan