+++ Changelog.md
@@ -5,10 +5,16 @@ Changelog
 
 `.env`:
 
+#### v17.10 rc4
+
+Changes:
+* Move `gogs` to `khs1994-docker/ci`
+* Move `ci` to `khs1994-docker/ci/webhooks`
+
 #### v17.10 rc3
 
 Changes:
-* Merge `arm32v7` `arm64v8` Compose file into one `docker-compose.arm.yml` 
+* Merge `arm32v7` `arm64v8` Compose file into one `docker-compose.arm.yml`
 
 #### v17.10 rc2